Split htcondor_service.py and its test file.
