+++ doc/changes/DM-52552.misc.rst
@@ -1,0 +1,1 @@
+Divided htcondor_service.py and test_htcondor_service.py into smaller files.
